+++ dockerfile
@@ -1,7 +1,8 @@
 FROM python:3.7
 # ADD ./crontask /etc/cron.d/crontask
-# RUN apt-get update \
-#     && apt-get install -y --no-install-recommends cron git wget vim \
+RUN pip install -U Flask \
+#     && apt-get update \
+#     && apt-get install -y --no-install-recommends cron git wget \
 #     && apt autoremove -y \
 #     && apt-get clean \
 #     && rm -rf /var/lib/apt/lists/* \
@@ -11,3 +12,4 @@ FROM python:3.7
 #     && echo "service cron start" >> /home/boot.sh \
 #     && echo "\n" >> /home/boot.sh \
 #     && echo "/bin/bash" >> /home/boot.sh
+WORKDIR /srv/iptv
